Recorded evidence of strange time degradation
When adding a new public function to `line.rs`, if the function is long
enough, it seems to disable the inlining of some nom methods, resulting
in a 20% degradation in process time via cargo bench. I've determined
this via flamegraph captures of before and after the change which show
in the faster time that the calls to nom are missing (implying that they
were inlined by the compiler, at least that's my interpretation). I'm
not sure why this is the case.