Spike on different looping strategies.
Seems that up_to_end_1 is the best for grabbing the slice up until the
end of the line. This will be used for slicing the HLS line when storing
a reference for writing.

Also double-checked the take until end of bytes method and seems fine
the way it is.