Reader from BufRead was too slow
Validating on each `fill_buf` call on the BufRead that the entire buffer
is a valid UTF-8 string is extremely wasteful; especially since we were
validating the same buffer over and over again.

I've made a change to check whether the contents of the BufRead are
valid UTF-8 on construction and then do the UTF-8 conversion unchecked
later on. This is NOT good enough because I can only check whet exists
in the buffer at point of construction, but more could get added later,
so this code is unsafe. I also changed to borrow from the BufRead
directly and only consume on the subsequent reads... I think that isn't
actually too crazy because you can only get to the BufRead by calling
into_inner at which point the Reader is consumed.

To be honest, I think I'm going to make a new commit to remove support
for BufRead until I change to the parser working with &[u8] instead of
&str.