Remove support for from_reader
I'm really not happy with the support I provided for parsing from a
Reader. Fundamentally it didn't make all that much sense and I think was
a little flawed anyway. The issue being that the reader is intended to
be reading data into a provided buffer, but as things stand, I don't
have a buffer to provide it. The trick with BufRead was to piggy-back
off of its internal buffer, but then the HLS line is only valid until
you need to read the next line, because you'll need to mutably borrow
the internal buffer again (which can't happen with an active reference
out somewhere). This _can_ be useful, for example, if we're immediately
writing the inner tag; however, I'm not confident that my usage of the
BufRead was correct enough to commit to supporting it (it feels strange
to rely on the fill_buf method, and not handle the interrupted signal,
and also only consume on the next read).

I may come back to add support for BufRead in the future. An example of
an alternative approach is with quick-xml, that requires the user of the
library to provide a String / Vec<u8> as a buffer to read data into and
then they parse from that buffer. This will have the same lifetime
problems I mentioned, but it seems at least like a more clear approach.