Update errors to provide problem line in response
This change allows anyone using the line parser to get back the input
when there is an error (which seems useful), but also, moves the reader
on to the next line in the case of a line error. This is important,
because before a line error would kill the whole reader (it would be
unusable after it hit an error because it would continuously try and
parse the same line and receive the same error). Now, if there is a
problematic line, then the user can be informed about it, and the reader
can move on to the next line and continue to be used until the end of
input.