Store enumerated as Cow str but reveal strong type
I think this is much more ergonomical and I doubt there is much cost to
doing so. Instead of storing as Enumerated<T> wherever it is used, we
continue to store as Cow<str>, but convert to Enumerated<T> on the get
methods. On each of the set methods we also continue to accept
impl Into<Cow<str>>, but Enumerated<T> implements Into<Cow<str>> so the
user can use Enumerated<T> or T directly.