Support for enumerated strings in EXT-X-MEDIA
This ended up being quite complicated, because I wanted to support the
new definitions of BED-n and DOF-n for CHANNELS, within an enumerated
string list of "special identifiers". Perhaps I could've created a new
type for this special case, but in the end, decided that perhaps it's
just better to allow the string list to be more flexible in this sense.
The problem being that, because these cases contain an associated (non
static) value, I could no longer assume T can always provide an as_str
method. So, to avoid a to_string (and always allocating) I made an
as_cow so that I could provide String only when needed.