Exclude .github and benches from package
Running cargo package before gives:
> Packaged 58 files, 1.6MiB (185.4KiB compressed)

Running cargo package after gives:
> Packaged 54 files, 572.6KiB (93.1KiB compressed)

The long_media_playlist.m3u8 used for benchmarking is 1.0M so this has a
major contribution to package size. I'm not sure if it is normal to
exclude benches from the package (for crates.io); however, I can see
that the memchr dependency does the same, and I believe that the author
of that library is a well known and respected member of the Rust
community, so I have trust that this must be OK to do. Perhaps I need to
do a little more research to be sure, but for now, this seems to make
sense to me.