Revamped how CustomTag works
There were several issues with how CustomTag was working previously:
* When writing, the custom tag output line was always recalculated, even
  if no mutation occurred.
* Providing a value for writing was not trivial due to the non-owned
  types required via SemiParsedTag / ParsedAttributeValue. This was more
  a wider problem that the library didn't really support writing custom
  data (constructed by the user rather than parsed by the library).

This is a significant change to how CustomTag behaves now. We always
wrap it in CustomTagAccess, which allows us to know if there was a
mutable borrow at some stage, and we use this to decide on whether we
need to recalculate the output line when writing. We've also introduced
mutable equivalents of SemiParsedTag and ParsedAttributeValue that hold
Cow-like values rather than ref-like values directly. This allows the
user to construct their own values if needed (and not have to rely on
storing parsed values in the TryFrom implementation).