Started implementing better builder syntax
The following video introduced me to the Sguaba library:
https://www.youtube.com/watch?v=EL7Au1tzNxE&pp=ygUMam9uIGdqZW5nc2V0

There I found this nice pattern of builder syntax with Rust that uses
generics to keep track of mandatory parameters that have or have not
been set, and only provide the "finish" method once all the
implementations of the generic patterns are set as the representation
used to indicate the mandatory parameter has been set.

This is a much nicer syntax because it allows all parameters in the
builder to be named, it looks like a true builder (no parameters on the
new method), and it enforces type safety on ensuring the mandatory
parameters have been set. I really like this approach.