feat!: new tag value abstraction
Added a new abstraction for reasoning about tag values. It is just a
wrapper around the raw input data and it provides a collection of
methods for parsing different data types from the data. The inention is
that the methods will be called when there is better context over what
the expected value is given the tag name. Still plenty left to do.

BREAKING CHANGE: removed the associated values from decimal resolution
parsing error (because there are multiple types of number failures now
and I don't think it adds too much information anyway).