feat: decimal resolution added for attribute value
Added the DecimalResolution case to the WritableAttributeValue as an
extra convenience (and to match the precedent set by WritableTagValue
that maps all the `try_as` methods to cases). Also added some extra
warning documentation around the usage of the "escape hatch" cases of
the writable types (not to introduce invalid characters as the lib does
not validate against it).