fix: delta update bench add skip only if any skip
Previously we were adding the skip tag unconditionally; however, if the
playlist was small enough such that no tags were skipped, then we would
add a skip tag with `SKIPPED-SEGMENTS=0`. This seems wrong and so has
been corrected so that the tag is not inserted if the removed count is
not greater than 0.