feat!: change lib name to quick-m3u8
The m3u8 name is taken on crates.io, so changing the lib name now in
preparation for publishing. It would be sad if this name also got taken
before I can get to publish, but lib names on crates.io are first come,
first served, and there is no author namespacing, so would just have to
move on with another name.

Ultimately, it's actually the "package name" that needs to be unique on
crates.io, and you can set the lib name to anything (and so can your
users). So I _could_ set the lib name to stay with `m3u8` and just
update the package name to something unique for crates.io. The drawback
of this approach is that it hurts discoverability of the package (not
that I really care much about that, this has mainly just been cathartic
to write), and it also could cause more confusion when a user installs
(given that it is irregular to have a lib name that differs from the
package name). In any case, there is a lot of back and forth about the
topic on this thread:
https://internals.rust-lang.org/t/crates-io-first-come-first-served-for-plain-crate-names-why-and-how-should-it-be-changed/19014