feat!: make config less public for forward compat
Made the properties of `ParsingOptions` proviate so that it cannot be
constructed directly. Also removed the `new` method so that construction
must go through the builder. These two changes means that adding config
properties will be easier to do without creating breaking changes, and I
believe that the API doesn't really suffer, so this seems like a better
approach.