feat!: reorganize module structure
This is an attempt to hvae a more coherent module structure that a user
may find easier to understand. The problems I saw with the previous mod
layout was:
* The `line` mod didn't seem very connected.
* The differences between `unknown`, `known`, and `value` seem a little
  unnecessary, given that a lot of the types are used equally.
* Re-using names between modules makes for awkward call-site usage (the
  main culprit being `Tag` that was defined in `known`, `unknown`, and
  `hls`, though the `hls` definition has stayed, but the others are
  renamed).
* The custom parsing methods seemed to be quite prominently exposed, but
  we don't really want users using them directly.

So the changes I looked for are:
* Moving most things to the top level of `tag` (so users don't know
  about the internal file organization).
* Gave unique names for the different types of `Tag`.
* Moved the custom parsing methods to be deeper in the module tree and
  clearly marked as "custom".

The drawback I don't like is that I've renamed the `tag` module used
internally as `tag_internal`, despite the fact that it's mostly public.
As of now, I can't think of a better name, but I may sleep on it and see
if something comes to me.