feat: added ordered attribute list for tag value
The ordered attribute list is a Vec instead of a HashMap. The HashMap
approach now uses the Vec method and converts to HashMap. Benchmarks
don't show that this is noticeably worse (and I will look at changing
to using Vec directly to see if things get any better, perhaps not
having to hash could be a benefit, but we'll see).

The change is primarily introduced because in my [other project] I need
to split attribute list tags into components, because I am using them to
build HTML markup, and I provide links to other content for some values
(e.g. URI on some tags, or SCTE35-OUT, etc.). I can achieve this with
the HashMap approach, but then the tag output for the viewer has the
order of attributes changed, and so the "manifest viewer" does not
reflect the true state of the playlist that was read. Introducing this
method will allow me to maintain the order of the attributes in the list
and perhaps this is the way it should've been done from the beginning.
But, I'm introducing this as a new method, so that this isn't a breaking
change.

[other project]: https://github.com/theRealRobG/hls-manifest-viewer