feat!: Daterange::start_date now optional
As of draft 18 it was clarified that the only mandatory attribute of the
EXT-X-DATERANGE tag is the ID. This is because subsequent daterange tags
with the same ID inherit the prvious attribute value definitions, and so
only the first tag instance must include a START-DATE. Given that tags
are all parsed in isolation (no state carried from previous line
parsings), we must consider the start_date optional now.

In the future, if we introduce a playlist abstraction, then we may be
able to provide a non-optional start date for a daterange.

BREAKING CHANGE: The type for start date on daterange has changed from
DateTime to Option<DateTime>.