feat!: consolidate decimal resolution parsing
There were two different parsing methods for DecimalResolution, one when
parsed from an unquoted string attribute value, and a separate method
for the `TryFrom<&str>` implementation. This change consolidates to one
parsing method by adding a new `TryFrom<&[u8]>` implementation which all
the other methods use internally.

BREAKING CHANGE: The `DecimalResolutionParseError` had a case removed as
it was no longer used. The `MissingHeight` error was renamed to
`MissingSeparator`, to better indicate the problem, which is that the
`x` separator is missing. The breaking change is introduced now as we
already have a few pending breaking changes for the next version.